MetaTempo integration is done for the web application (BimodLab).

+++ apps/metatempo/views/board/metatempo_purge_logs_views.py
@@ -19,7 +19,9 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import redirect
 from django.views import View
 
+from apps.core.user_permissions.permission_manager import UserPermissionManager
 from apps.metatempo.models import MetaTempoMemberLog, MetaTempoMemberLogDaily, MetaTempoProjectOverallLog
+from apps.user_permissions.utils import PermissionNames
 
 
 class MetaTempoView_PurgeLogs(LoginRequiredMixin, View):
@@ -28,6 +30,15 @@ class MetaTempoView_PurgeLogs(LoginRequiredMixin, View):
 
     def post(self, request, *args, **kwargs):
         connection_id = kwargs.get('connection_id')
+
+        ##############################
+        # PERMISSION CHECK FOR - DELETE_METATEMPO_CONNECTION
+        if not UserPermissionManager.is_authorized(user=self.request.user,
+                                                   operation=PermissionNames.DELETE_METATEMPO_CONNECTION):
+            messages.error(self.request, "You do not have permission to delete MetaTempo Connection logs.")
+            return redirect('metatempo:main_board', connection_id=connection_id)
+        ##############################
+
         try:
             individual_logs_deleted, _ = MetaTempoMemberLog.objects.filter(
                 metatempo_connection_id=connection_id).delete()
